requirements.txt
flask
easyocr
pillow
torch
opencv-python-headless